Copied project from qidtrace2 to qidtrace_nosave
Copied files from qidtrace2 directory to qidtrace_nosave directory,
including the git history (three commits were made to the qidtrace2 project);
so this is similar to renaming the project.

This first commit in the new project dir includes only the README
and INTRO files, w/ small text changes related to this rename.