qidtrace: push anon hash ref onto @_trailing array.
The anon hash contains the current line and its line number,
exactly what is saved to the @_leading array.  Previously,
it was holding only the line of data itself, a remnant from
prior work.

Also removed extraneous print subs used for debugging, and
cleaned up the file in other ways.